+++ Runtime/CompilerServices/CompilerServices.cs
@@ -1,7 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-
-namespace System.Runtime.CompilerServices
-{
-    public static class IsExternalInit { }
-}