+++ Runtime/Mathematics/Vector2IntExtensions.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+namespace Jeomseon.Mathematics
+{
+    public static class Vector2IntExtensions
+    {
+        public static Vector2Int Abs(this Vector2Int vec) => new(
+            Mathf.Abs(vec.x),
+            Mathf.Abs(vec.y));
+    }
+}