+++ Runtime/Mathematics/Vector3IntExtensions.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace Jeomseon.Mathematics
+{
+    public static class Vector3IntExtensions
+    {
+        public static Vector3Int Abs(this Vector3Int vec) => new(
+            Mathf.Abs(vec.x),
+            Mathf.Abs(vec.y),
+            Mathf.Abs(vec.z));
+    }
+}